[REFACTOR] refactor to accomodate special round in specific components; remove timeframe store; create useTimeframe hook
