+++ components/Pages/Launchpad/ProjectDetails/SpecialRound/index.tsx
@@ -1,13 +1,16 @@
 import Image from "next/image";
 import Icon from "@/components/Icon";
-import RoundContractInteractions from "./RoundContractInteractions";
+import RoundContractInteractions from "../LauchpadBanner/RoundContractInteractions";
 import Link from "next/link";
 import { Project } from "@/types";
 import { useMemo } from "react";
 import { useContractRead } from "wagmi";
 import { formatDisplayedBalance, getRoundAbi } from "@/utils";
+import { format } from "date-fns";
+import TimeframeDropdown from "../SpecialRound/TimeframeDropdown";
+import { useTimeframe } from "@/hooks/useTimeframe";
 
-export default function ProjectPageBanner({ project }: { project: Project }) {
+export default function SpecialRound({ project }: { project: Project }) {
   const activeRound = useMemo(() => {
     const active = project.rounds.find((round) => {
       return (
@@ -20,6 +23,7 @@ export default function ProjectPageBanner({ project }: { project: Project }) {
     });
     return active || next || project.rounds[0];
   }, [project]);
+  const { hasTimeframe } = useTimeframe(activeRound);
 
   const { data: roundData } = useContractRead({
     address: activeRound.address,
@@ -102,6 +106,38 @@ export default function ProjectPageBanner({ project }: { project: Project }) {
 
           <p className="text-secondary text-body-14">{project.description}</p>
         </div>
+
+        {hasTimeframe && (
+          <div className="mb-10 w-full rounded-lg bg-surface-soft flex gap-4 desktop:gap-20 p-4 flex-col desktop:flex-row tablet:flex-row gap-x-10">
+            <div className="flex justify-between desktop:gap-20 tablet:gap-x-10">
+              <div className="flex flex-col">
+                <p className="text-lg text-secondary font-medium">
+                  Available From
+                </p>
+                <p className="text-lg mt-2">
+                  {format(new Date(activeRound.start), "dd/MM/yyyy")}
+                </p>
+              </div>
+              <div className="flex flex-col">
+                <p className="text-lg text-secondary font-medium">To</p>
+                <p className="text-lg mt-2">
+                  {format(new Date(activeRound.end), "dd/MM/yyyy")}
+                </p>
+              </div>
+            </div>
+            <div className="flex flex-col grow gap-2">
+              <div className="flex justify-between items-baseline">
+                <p className="text-lg text-primary font-bold">Timeframes</p>
+                <p className="text-sm text-tertiary font-thin">
+                  Available at these hours everyday
+                </p>
+              </div>
+              <div className="text-body-16 text-secondary font-medium">
+                <TimeframeDropdown round={activeRound} />
+              </div>
+            </div>
+          </div>
+        )}
 
         <RoundContractInteractions
           round={activeRound}
